Cleanup old metrics test
Remove Go 1.18 test now that 1.19 is the minimum supported version.
